GH-3: Enhance the performance for 5 times  - Improve the `Ffmpeg` to not depend on member properties while using the instance in multiple threads  - Run worker in separate processes and get frames concurrently (using threads)
